requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==3.0.1
django-pandas==0.6.6
docx2pdf==0.1.8
et-xmlfile==1.1.0
filelock==3.8.0
fpdf==1.7.2
huggingface-hub==0.10.1
idna==3.4
joblib==1.2.0
lxml==4.9.1
MarkupPy==1.14
nltk==3.7
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
PyDispatcher==2.0.5
pyparsing==3.0.9
PyPDF2==2.11.1
PyPDF4==1.27.0
pyreadline==2.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
PyYAML==6.0
queuelib==1.5.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tabulate==0.8.9
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
PyMuPDF==1.21.1
django-quill-editor==0.1.40
psycopg2==2.9.6
openai==0.27.4
Unidecode==1.3.6
doc2pdf==0.2.0
wget==3.2
aspose-words==23.3.0
docxcompose==1.4.0
azure-storage-blob == 12.15.0
# pywin32==302
# openai - should be installed
# django_quill - should be installed